requirements.txt
flask==3.1.1
opencv-python==4.11.0.86
ultralytics==8.3.127
numpy==1.26.4
python-dotenv==1.1.0
flask-cors==6.0.1
PyJWT==2.10.1
tensorflow==2.19.0
deepface==0.0.93
flask_sqlalchemy==3.1.1